requirements.txt
-e .
black==23.1.0
coverage==7.2.2
GitPython==3.1.31
mypy==1.1.1
pdoc==13.0.0
pip==23.0.1
pre-commit==3.1.1
pytest==7.2.2
pytest-asyncio==0.20.3
pytest-cov==4.0.0
PyGithub==1.58.0
PyYAML==6.0
ruff==0.0.256
types-PyYAML==6.0.12.8
typing-extensions==4.5.0
wheel==0.40.0
yamllint==1.29.0
